CODE/requirements.txt
Werkzeug==2.2.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
SQLAlchemy==2.0.6
WTForms==3.0.1
zipp==3.15.0
matplotlib==3.8.4
celery==5.3.5
redis==5.0.1
Flask-Caching==2.1.0